Use new json data
In a previous branch, I switched to using the json supplied by adunis, instead of crawling the html. This new data is now used in this branch.
A mapping of 2 old module names (FunProg to FP and BAI14 to BAI21) was required. The module names in the url get updated as well. We should be able to remove that mapping after some months.
We had to introduce a hirarchical priority for the category colors,
since some modules can belong to multiple categories.
I also had to update some dependencies to get it running at all.